Change match_id type from u64 to u32 across the codebase
Updated the match_id type to align with expected data size, simplifying logic and ensuring consistency. Adjusted associated functions, structs, and type casts accordingly to reflect this change.

+++ salt-scraper/src/models.rs
@@ -3,13 +3,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Row, Deserialize, PartialEq, Eq, Hash, Clone)]
 pub struct MatchIdQueryResult {
-    pub(crate) match_id: u64,
+    pub(crate) match_id: u32,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct MatchSalt {
     pub(crate) cluster_id: u32,
-    pub(crate) match_id: u64,
+    pub(crate) match_id: u32,
     pub(crate) metadata_salt: u32,
     pub(crate) replay_salt: u32,
     pub(crate) username: Option<String>,
